Conflicts:
	app/views/changeset/list.html.erb

+++ app/controllers/user_controller.rb
@@ -1,20 +1,22 @@
 class UserController < ApplicationController
   layout :choose_layout
 
-  skip_before_filter :verify_authenticity_token, :only => [:api_details, :api_gpx_files]
+  skip_before_filter :verify_authenticity_token, :only => [:api_read, :api_details, :api_gpx_files]
   before_filter :disable_terms_redirect, :only => [:terms, :save, :logout, :api_details]
   before_filter :authorize, :only => [:api_details, :api_gpx_files]
-  before_filter :authorize_web, :except => [:api_details, :api_gpx_files]
-  before_filter :set_locale, :except => [:api_details, :api_gpx_files]
+  before_filter :authorize_web, :except => [:api_read, :api_details, :api_gpx_files]
+  before_filter :set_locale, :except => [:api_read, :api_details, :api_gpx_files]
   before_filter :require_user, :only => [:account, :go_public, :make_friend, :remove_friend]
-  before_filter :check_database_readable, :except => [:login, :api_details, :api_gpx_files]
+  before_filter :check_database_readable, :except => [:login, :api_read, :api_details, :api_gpx_files]
   before_filter :check_database_writable, :only => [:new, :account, :confirm, :confirm_email, :lost_password, :reset_password, :go_public, :make_friend, :remove_friend]
-  before_filter :check_api_readable, :only => [:api_details, :api_gpx_files]
+  before_filter :check_api_readable, :only => [:api_read, :api_details, :api_gpx_files]
   before_filter :require_allow_read_prefs, :only => [:api_details]
   before_filter :require_allow_read_gpx, :only => [:api_gpx_files]
   before_filter :require_cookies, :only => [:login, :confirm]
   before_filter :require_administrator, :only => [:set_status, :delete, :list]
-  before_filter :lookup_this_user, :only => [:set_status, :delete]
+  around_filter :api_call_handle_error, :only => [:api_read, :api_details, :api_gpx_files]
+  before_filter :lookup_user_by_id, :only => [:api_read]
+  before_filter :lookup_user_by_name, :only => [:set_status, :delete]
 
   cache_sweeper :user_sweeper, :only => [:account, :set_status, :delete]
 
@@ -30,12 +32,14 @@ class UserController < ApplicationController
       # the open_id_authentication function
       @user = session.delete(:new_user)
 
-      openid_verify(nil, @user) do |user|
+      openid_verify(nil, @user) do |user,verified_email|
+        user.status = "active" if user.email == verified_email
       end
 
       if @user.openid_url.nil? or @user.invalid?
         render :action => 'new'
       else
+        session[:new_user] = @user
         render :action => 'terms'
       end
     elsif params[:user] and Acl.no_account_creation(request.remote_ip, params[:user][:email].split("@").last)
@@ -49,11 +53,13 @@ class UserController < ApplicationController
       if params[:user] and params[:user][:openid_url] and @user.pass_crypt.empty?
         # We are creating an account with OpenID and no password
         # was specified so create a random one
-        @user.pass_crypt = SecureRandom.base64(16) 
-        @user.pass_crypt_confirmation = @user.pass_crypt 
+        @user.pass_crypt = SecureRandom.base64(16)
+        @user.pass_crypt_confirmation = @user.pass_crypt
       end
 
       if @user
+        @user.status = "pending"
+
         if @user.invalid?
           if @user.new_record?
             # Something is wrong with a new user, so rerender the form
@@ -70,6 +76,9 @@ class UserController < ApplicationController
           # Verify OpenID before moving on
           session[:new_user] = @user
           openid_verify(params[:user][:openid_url], @user)
+        elsif @user.new_record?
+          # Save the user record
+          session[:new_user] = @user
         end
       else
         # Not logged in, so redirect to the login page
@@ -112,28 +121,40 @@ class UserController < ApplicationController
       else
         redirect_to :action => :account, :display_name => @user.display_name
       end
-    elsif Acl.no_account_creation(request.remote_ip, params[:user][:email].split("@").last)
-      render :action => 'blocked'
     else
-      @user = User.new(params[:user])
+      @user = session.delete(:new_user)
 
-      @user.status = "pending"
-      @user.data_public = true
-      @user.description = "" if @user.description.nil?
-      @user.creation_ip = request.remote_ip
-      @user.languages = request.user_preferred_languages
-      @user.terms_agreed = Time.now.getutc
-      @user.terms_seen = true
-      @user.openid_url = nil if @user.openid_url and @user.openid_url.empty?
-      
-      if @user.save
-        flash[:piwik_goal] = PIWIK_SIGNUP_GOAL if defined?(PIWIK_SIGNUP_GOAL)
-        flash[:notice] = t 'user.new.flash create success message', :email => @user.email
-        Notifier.signup_confirm(@user, @user.tokens.create(:referer => session.delete(:referer))).deliver
-        session[:token] = @user.tokens.create.token
-        redirect_to :action => 'login', :referer => params[:referer]
+      if Acl.no_account_creation(request.remote_ip, @user.email.split("@").last)
+        render :action => 'blocked'
       else
-        render :action => 'new', :referer => params[:referer]
+        @user.data_public = true
+        @user.description = "" if @user.description.nil?
+        @user.creation_ip = request.remote_ip
+        @user.languages = request.user_preferred_languages
+        @user.terms_agreed = Time.now.getutc
+        @user.terms_seen = true
+        @user.openid_url = nil if @user.openid_url and @user.openid_url.empty?
+
+        if @user.save
+          flash[:piwik_goal] = PIWIK_SIGNUP_GOAL if defined?(PIWIK_SIGNUP_GOAL)
+
+          if @user.status == "active"
+            flash[:notice] = t 'user.new.flash welcome', :email => @user.email
+
+            Notifier.signup_confirm(@user, nil).deliver
+
+            successful_login(@user)
+          else
+            flash[:notice] = t 'user.new.flash create success message', :email => @user.email
+            session[:token] = @user.tokens.create.token
+
+            Notifier.signup_confirm(@user, @user.tokens.create(:referer => session.delete(:referer))).deliver
+
+            redirect_to :action => 'login', :referer => params[:referer]
+          end
+        else
+          render :action => 'new', :referer => params[:referer]
+        end
       end
     end
   end
@@ -143,55 +164,24 @@ class UserController < ApplicationController
     @tokens = @user.oauth_tokens.authorized
 
     if params[:user] and params[:user][:display_name] and params[:user][:description]
-      @user.display_name = params[:user][:display_name]
-      @user.new_email = params[:user][:new_email]
-
-      if params[:user][:pass_crypt].length > 0 or params[:user][:pass_crypt_confirmation].length > 0
-        @user.pass_crypt = params[:user][:pass_crypt]
-        @user.pass_crypt_confirmation = params[:user][:pass_crypt_confirmation]
-      end
-
-      if params[:user][:description] != @user.description
-        @user.description = params[:user][:description]
-        @user.description_format = "markdown"
-      end
-
-      @user.languages = params[:user][:languages].split(",")
-
-      case params[:image_action]
-        when "new" then @user.image = params[:user][:image]
-        when "delete" then @user.image = nil
-      end
-
-      @user.home_lat = params[:user][:home_lat]
-      @user.home_lon = params[:user][:home_lon]
-
-      if params[:user][:preferred_editor] == "default"
-        @user.preferred_editor = nil
-      else
-        @user.preferred_editor = params[:user][:preferred_editor]
-      end
-
-      @user.openid_url = nil if params[:user][:openid_url].blank?
-
       if params[:user][:openid_url] and
          params[:user][:openid_url].length > 0 and
          params[:user][:openid_url] != @user.openid_url
         # If the OpenID has changed, we want to check that it is a
         # valid OpenID and one the user has control over before saving
         # it as a password equivalent for the user.
-        session[:new_user] = @user
+        session[:new_user_settings] = params
         openid_verify(params[:user][:openid_url], @user)
       else
-        update_user(@user)
+        update_user(@user, params)
       end
     elsif using_open_id?
       # The redirect from the OpenID provider reenters here
       # again and we need to pass the parameters through to
       # the open_id_authentication function
-      @user = session.delete(:new_user)
+      settings = session.delete(:new_user_settings)
       openid_verify(nil, @user) do |user|
-        update_user(user)
+        update_user(user, settings)
       end
     end
   end
@@ -290,10 +280,6 @@ class UserController < ApplicationController
       else
         password_authentication(params[:username], params[:password])
       end
-    elsif params[:notice]
-      flash.now[:notice] = t "user.login.notice_#{params[:notice]}"
-    elsif flash[:notice].nil?
-      flash.now[:notice] =  t 'user.login.notice'
     end
   end
 
@@ -332,7 +318,7 @@ class UserController < ApplicationController
           referer = token.referer
           token.destroy
 
-          if session[:token] 
+          if session[:token]
             token = UserToken.find_by_token(session[:token])
             session.delete(:token)
           else
@@ -408,6 +394,15 @@ class UserController < ApplicationController
     end
   end
 
+  def api_read
+    render :nothing => true, :status => :gone unless @this_user.visible?
+  end
+
+  def api_details
+    @this_user = @user
+    render :action => :api_read
+  end
+
   def api_gpx_files
     doc = OSM::API.new.get_xml_doc
     @user.traces.each do |trace|
@@ -428,47 +423,55 @@ class UserController < ApplicationController
   end
 
   def make_friend
-    if params[:display_name]
-      name = params[:display_name]
-      new_friend = User.active.where(:display_name => name).first
-      friend = Friend.new
-      friend.user_id = @user.id
-      friend.friend_user_id = new_friend.id
-      unless @user.is_friends_with?(new_friend)
-        if friend.save
-          flash[:notice] = t 'user.make_friend.success', :name => name
-          Notifier.friend_notification(friend).deliver
-        else
-          friend.add_error(t('user.make_friend.failed', :name => name))
-        end
-      else
-        flash[:warning] = t 'user.make_friend.already_a_friend', :name => name
-      end
+    @new_friend = User.find_by_display_name(params[:display_name])
 
-      if params[:referer]
-        redirect_to params[:referer]
-      else
-        redirect_to :controller => 'user', :action => 'view'
+    if @new_friend
+      if request.post?
+        friend = Friend.new
+        friend.user_id = @user.id
+        friend.friend_user_id = @new_friend.id
+        unless @user.is_friends_with?(@new_friend)
+          if friend.save
+            flash[:notice] = t 'user.make_friend.success', :name => @new_friend.display_name
+            Notifier.friend_notification(friend).deliver
+          else
+            friend.add_error(t('user.make_friend.failed', :name => @new_friend.display_name))
+          end
+        else
+          flash[:warning] = t 'user.make_friend.already_a_friend', :name => @new_friend.display_name
+        end
+
+        if params[:referer]
+          redirect_to params[:referer]
+        else
+          redirect_to :controller => 'user', :action => 'view'
+        end
       end
+    else
+      render_unknown_user params[:display_name]
     end
   end
 
   def remove_friend
-    if params[:display_name]
-      name = params[:display_name]
-      friend = User.active.where(:display_name => name).first
-      if @user.is_friends_with?(friend)
-        Friend.delete_all "user_id = #{@user.id} AND friend_user_id = #{friend.id}"
-        flash[:notice] = t 'user.remove_friend.success', :name => friend.display_name
-      else
-        flash[:error] = t 'user.remove_friend.not_a_friend', :name => friend.display_name
-      end
+    @friend = User.find_by_display_name(params[:display_name])
 
-      if params[:referer]
-        redirect_to params[:referer]
-      else
-        redirect_to :controller => 'user', :action => 'view'
+    if @friend
+      if request.post?
+        if @user.is_friends_with?(@friend)
+          Friend.delete_all "user_id = #{@user.id} AND friend_user_id = #{@friend.id}"
+          flash[:notice] = t 'user.remove_friend.success', :name => @friend.display_name
+        else
+          flash[:error] = t 'user.remove_friend.not_a_friend', :name => @friend.display_name
+        end
+
+        if params[:referer]
+          redirect_to params[:referer]
+        else
+          redirect_to :controller => 'user', :action => 'view'
+        end
       end
+    else
+      render_unknown_user params[:display_name]
     end
   end
 
@@ -548,7 +551,7 @@ private
         if user = User.find_by_openid_url(identity_url)
           case user.status
             when "pending" then
-              failed_login t('user.login.account not active')
+              failed_login t('user.login.account not active', :reconfirm => url_for(:action => 'confirm_resend', :display_name => user.display_name))
             when "active", "confirmed" then
               successful_login(user)
             when "suspended" then
@@ -565,8 +568,9 @@ private
           # to the create account page with username and email filled
           # in if they have been given by the OpenID provider through
           # the simple registration protocol.
-          nickname = sreg["nickname"] || ax["http://axschema.org/namePerson/friendly"]
-          email = sreg["email"] || ax["http://axschema.org/contact/email"]
+          nickname = sreg["nickname"] || ax["http://axschema.org/namePerson/friendly"].first
+          email = sreg["email"] || ax["http://axschema.org/contact/email"].first
+
           redirect_to :controller => 'user', :action => 'new', :nickname => nickname, :email => email, :openid => identity_url
         end
       elsif result.missing?
@@ -584,8 +588,18 @@ private
   def openid_verify(openid_url, user)
     user.openid_url = openid_url
 
-    authenticate_with_open_id(openid_expand_url(openid_url), :method => :get) do |result, identity_url|
+    authenticate_with_open_id(openid_expand_url(openid_url), :method => :get, :required => [:email, "http://axschema.org/contact/email"]) do |result, identity_url, sreg, ax|
       if result.successful?
+        # Do we trust the emails this provider returns?
+        if openid_email_verified(identity_url)
+          # Guard against not getting any extension data
+          sreg = Hash.new if sreg.nil?
+          ax = Hash.new if ax.nil?
+
+          # Get the verified email
+          verified_email = sreg["email"] || ax["http://axschema.org/contact/email"].first
+        end
+
         # We need to use the openid url passed back from the OpenID provider
         # rather than the one supplied by the user, as these can be different.
         #
@@ -593,7 +607,7 @@ private
         # than a user specific url. Only once it comes back from the provider
         # provider do we know the unique address for the user.
         user.openid_url = identity_url
-        yield user
+        yield user, verified_email
       elsif result.missing?
         flash.now[:error] = t 'user.login.openid missing provider'
       elsif result.invalid?
@@ -619,7 +633,15 @@ private
     else
       return openid_url
     end
-  end  
+  end
+
+  ##
+  # check if we trust an OpenID provider to return a verified
+  # email, so that we can skpi verifying it ourselves
+  def openid_email_verified(openid_url)
+    openid_url.match(/https:\/\/www.google.com\/accounts\/o8\/id?(.*)/) or
+    openid_url.match(/https:\/\/me.yahoo.com\/(.*)/)
+  end
 
   ##
   # process a successful login
@@ -662,7 +684,45 @@ private
 
   ##
   # update a user's details
-  def update_user(user)
+  def update_user(user, params)
+    user.display_name = params[:user][:display_name]
+    user.new_email = params[:user][:new_email]
+
+    if params[:user][:pass_crypt].length > 0 or params[:user][:pass_crypt_confirmation].length > 0
+      user.pass_crypt = params[:user][:pass_crypt]
+      user.pass_crypt_confirmation = params[:user][:pass_crypt_confirmation]
+    end
+
+    if params[:user][:description] != user.description
+      user.description = params[:user][:description]
+      user.description_format = "markdown"
+    end
+
+    user.languages = params[:user][:languages].split(",")
+
+    case params[:image_action]
+    when "new" then
+      user.image = params[:user][:image]
+      user.image_use_gravatar = false
+    when "delete" then
+      user.image = nil
+      user.image_use_gravatar = false
+    when "gravatar" then
+      user.image = nil
+      user.image_use_gravatar = true
+    end
+
+    user.home_lat = params[:user][:home_lat]
+    user.home_lon = params[:user][:home_lon]
+
+    if params[:user][:preferred_editor] == "default"
+      user.preferred_editor = nil
+    else
+      user.preferred_editor = params[:user][:preferred_editor]
+    end
+
+    user.openid_url = nil if params[:user][:openid_url].blank?
+
     if user.save
       set_locale
 
@@ -710,7 +770,13 @@ private
 
   ##
   # ensure that there is a "this_user" instance variable
-  def lookup_this_user
+  def lookup_user_by_id
+    @this_user = User.find(params[:id])
+  end
+
+  ##
+  # ensure that there is a "this_user" instance variable
+  def lookup_user_by_name
     @this_user = User.find_by_display_name(params[:display_name])
   rescue ActiveRecord::RecordNotFound
     redirect_to :controller => 'user', :action => 'view', :display_name => params[:display_name] unless @this_user
@@ -734,7 +800,7 @@ private
   ##
   #
   def disable_terms_redirect
-    # this is necessary otherwise going to the user terms page, when 
+    # this is necessary otherwise going to the user terms page, when
     # having not agreed already would cause an infinite redirect loop.
     # it's .now so that this doesn't propagate to other pages.
     flash.now[:skip_terms] = true
